requirements/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
pytest>=7.4.3
uvicorn>=0.15.0,<0.16.0
python-dotenv>=1.0.0